Update checkout form as details are updated
Changed the hidden form from being generated on the server side to
be generated on the client side instead. Added event listeners to
update the form as user details are modified.

Added scope settings to allow branding configuration to be changed
at the store level. Updated front end code to request data in the
correct scope.

Changes should fix issues #10, #17, #18 and #22.