Restrict orders to specific countries
Added the "Payment from Specific Countries" field to the config,
restricted to only allow selction of AU and/or NZ. This is a
built in Magento field that hides the payment option if a country
outside of the selected values is set for shipping.
Added error checking to stop orders from outside the selected
countries. This is primarily for billing address, but also checks
the shipping country in the cases where the magento logic fails.
Improved error checking functionality and error display in the
checkout.

Fixes issue #8